Warns if datasets overlap, improved click command readability.
